[python-sdk] move tokenv1 code out of async client
just make the api a bit cleaner since this isn't the main token standard
any more
