+++ ecosystem/python/sdk/aptos_sdk/aptos_tokenv1_client.py
@@ -1,0 +1,294 @@
+# Copyright © Aptos Foundation
+# SPDX-License-Identifier: Apache-2.0
+
+from typing import Any
+
+from .account import Account
+from .account_address import AccountAddress
+from .async_client import ApiError, RestClient
+from .bcs import Serializer
+from .transactions import EntryFunction, TransactionArgument, TransactionPayload
+
+U64_MAX = 18446744073709551615
+
+
+class AptosTokenV1Client:
+    """A wrapper around reading and mutating AptosTokens also known as Token Objects"""
+
+    _client: RestClient
+
+    def __init__(self, client: RestClient):
+        self._client = client
+
+    async def create_collection(
+        self, account: Account, name: str, description: str, uri: str
+    ) -> str:
+        """Creates a new collection within the specified account"""
+
+        transaction_arguments = [
+            TransactionArgument(name, Serializer.str),
+            TransactionArgument(description, Serializer.str),
+            TransactionArgument(uri, Serializer.str),
+            TransactionArgument(U64_MAX, Serializer.u64),
+            TransactionArgument(
+                [False, False, False], Serializer.sequence_serializer(Serializer.bool)
+            ),
+        ]
+
+        payload = EntryFunction.natural(
+            "0x3::token",
+            "create_collection_script",
+            [],
+            transaction_arguments,
+        )
+
+        signed_transaction = await self._client.create_bcs_signed_transaction(
+            account, TransactionPayload(payload)
+        )
+        return await self._client.submit_bcs_transaction(signed_transaction)
+
+    async def create_token(
+        self,
+        account: Account,
+        collection_name: str,
+        name: str,
+        description: str,
+        supply: int,
+        uri: str,
+        royalty_points_per_million: int,
+    ) -> str:
+        transaction_arguments = [
+            TransactionArgument(collection_name, Serializer.str),
+            TransactionArgument(name, Serializer.str),
+            TransactionArgument(description, Serializer.str),
+            TransactionArgument(supply, Serializer.u64),
+            TransactionArgument(supply, Serializer.u64),
+            TransactionArgument(uri, Serializer.str),
+            TransactionArgument(account.address(), Serializer.struct),
+            # SDK assumes per million
+            TransactionArgument(1000000, Serializer.u64),
+            TransactionArgument(royalty_points_per_million, Serializer.u64),
+            TransactionArgument(
+                [False, False, False, False, False],
+                Serializer.sequence_serializer(Serializer.bool),
+            ),
+            TransactionArgument([], Serializer.sequence_serializer(Serializer.str)),
+            TransactionArgument(
+                [], Serializer.sequence_serializer(Serializer.to_bytes)
+            ),
+            TransactionArgument([], Serializer.sequence_serializer(Serializer.str)),
+        ]
+
+        payload = EntryFunction.natural(
+            "0x3::token",
+            "create_token_script",
+            [],
+            transaction_arguments,
+        )
+        signed_transaction = await self._client.create_bcs_signed_transaction(
+            account, TransactionPayload(payload)
+        )
+        return await self._client.submit_bcs_transaction(signed_transaction)
+
+    async def offer_token(
+        self,
+        account: Account,
+        receiver: AccountAddress,
+        creator: AccountAddress,
+        collection_name: str,
+        token_name: str,
+        property_version: int,
+        amount: int,
+    ) -> str:
+        transaction_arguments = [
+            TransactionArgument(receiver, Serializer.struct),
+            TransactionArgument(creator, Serializer.struct),
+            TransactionArgument(collection_name, Serializer.str),
+            TransactionArgument(token_name, Serializer.str),
+            TransactionArgument(property_version, Serializer.u64),
+            TransactionArgument(amount, Serializer.u64),
+        ]
+
+        payload = EntryFunction.natural(
+            "0x3::token_transfers",
+            "offer_script",
+            [],
+            transaction_arguments,
+        )
+        signed_transaction = await self._client.create_bcs_signed_transaction(
+            account, TransactionPayload(payload)
+        )
+        return await self._client.submit_bcs_transaction(signed_transaction)
+
+    async def claim_token(
+        self,
+        account: Account,
+        sender: AccountAddress,
+        creator: AccountAddress,
+        collection_name: str,
+        token_name: str,
+        property_version: int,
+    ) -> str:
+        transaction_arguments = [
+            TransactionArgument(sender, Serializer.struct),
+            TransactionArgument(creator, Serializer.struct),
+            TransactionArgument(collection_name, Serializer.str),
+            TransactionArgument(token_name, Serializer.str),
+            TransactionArgument(property_version, Serializer.u64),
+        ]
+
+        payload = EntryFunction.natural(
+            "0x3::token_transfers",
+            "claim_script",
+            [],
+            transaction_arguments,
+        )
+        signed_transaction = await self._client.create_bcs_signed_transaction(
+            account, TransactionPayload(payload)
+        )
+        return await self._client.submit_bcs_transaction(signed_transaction)
+
+    async def direct_transfer_token(
+        self,
+        sender: Account,
+        receiver: Account,
+        creators_address: AccountAddress,
+        collection_name: str,
+        token_name: str,
+        property_version: int,
+        amount: int,
+    ) -> str:
+        transaction_arguments = [
+            TransactionArgument(creators_address, Serializer.struct),
+            TransactionArgument(collection_name, Serializer.str),
+            TransactionArgument(token_name, Serializer.str),
+            TransactionArgument(property_version, Serializer.u64),
+            TransactionArgument(amount, Serializer.u64),
+        ]
+
+        payload = EntryFunction.natural(
+            "0x3::token",
+            "direct_transfer_script",
+            [],
+            transaction_arguments,
+        )
+
+        signed_transaction = await self._client.create_multi_agent_bcs_transaction(
+            sender,
+            [receiver],
+            TransactionPayload(payload),
+        )
+        return await self._client.submit_bcs_transaction(signed_transaction)
+
+    #
+    # Token accessors
+    #
+
+    async def get_token(
+        self,
+        owner: AccountAddress,
+        creator: AccountAddress,
+        collection_name: str,
+        token_name: str,
+        property_version: int,
+    ) -> Any:
+        resource = await self._client.account_resource(owner, "0x3::token::TokenStore")
+        token_store_handle = resource["data"]["tokens"]["handle"]
+
+        token_id = {
+            "token_data_id": {
+                "creator": str(creator),
+                "collection": collection_name,
+                "name": token_name,
+            },
+            "property_version": str(property_version),
+        }
+
+        try:
+            return await self._client.get_table_item(
+                token_store_handle,
+                "0x3::token::TokenId",
+                "0x3::token::Token",
+                token_id,
+            )
+        except ApiError as e:
+            if e.status_code == 404:
+                return {
+                    "id": token_id,
+                    "amount": "0",
+                }
+            raise
+
+    async def get_token_balance(
+        self,
+        owner: AccountAddress,
+        creator: AccountAddress,
+        collection_name: str,
+        token_name: str,
+        property_version: int,
+    ) -> str:
+        info = await self.get_token(
+            owner, creator, collection_name, token_name, property_version
+        )
+        return info["amount"]
+
+    async def get_token_data(
+        self,
+        creator: AccountAddress,
+        collection_name: str,
+        token_name: str,
+        property_version: int,
+    ) -> Any:
+        resource = await self._client.account_resource(
+            creator, "0x3::token::Collections"
+        )
+        token_data_handle = resource["data"]["token_data"]["handle"]
+
+        token_data_id = {
+            "creator": str(creator),
+            "collection": collection_name,
+            "name": token_name,
+        }
+
+        return await self._client.get_table_item(
+            token_data_handle,
+            "0x3::token::TokenDataId",
+            "0x3::token::TokenData",
+            token_data_id,
+        )  # <:!:read_token_data_table
+
+    async def get_collection(
+        self, creator: AccountAddress, collection_name: str
+    ) -> Any:
+        resource = await self._client.account_resource(
+            creator, "0x3::token::Collections"
+        )
+        token_data = resource["data"]["collection_data"]["handle"]
+
+        return await self._client.get_table_item(
+            token_data,
+            "0x1::string::String",
+            "0x3::token::CollectionData",
+            collection_name,
+        )
+
+    async def transfer_object(
+        self, owner: Account, object: AccountAddress, to: AccountAddress
+    ) -> str:
+        transaction_arguments = [
+            TransactionArgument(object, Serializer.struct),
+            TransactionArgument(to, Serializer.struct),
+        ]
+
+        payload = EntryFunction.natural(
+            "0x1::object",
+            "transfer_call",
+            [],
+            transaction_arguments,
+        )
+
+        signed_transaction = await self._client.create_bcs_signed_transaction(
+            owner,
+            TransactionPayload(payload),
+        )
+        return await self._client.submit_bcs_transaction(signed_transaction)
