+++ ecosystem/python/sdk/examples/simple_nft.py
@@ -5,15 +5,16 @@ import asyncio
 import json
 
 from aptos_sdk.account import Account
+from aptos_sdk.aptos_tokenv1_client import AptosTokenV1Client
 from aptos_sdk.async_client import FaucetClient, RestClient
 
 from .common import FAUCET_URL, NODE_URL
 
 
 async def main():
-    # :!:>section_1
     rest_client = RestClient(NODE_URL)
-    faucet_client = FaucetClient(FAUCET_URL, rest_client)  # <:!:section_1
+    faucet_client = FaucetClient(FAUCET_URL, rest_client)
+    token_client = AptosTokenV1Client(rest_client)
 
     # :!:>section_2
     alice = Account.generate()
@@ -42,13 +43,13 @@ async def main():
     print("\n=== Creating Collection and Token ===")
 
     # :!:>section_4
-    txn_hash = await rest_client.create_collection(
+    txn_hash = await token_client.create_collection(
         alice, collection_name, "Alice's simple collection", "https://aptos.dev"
     )  # <:!:section_4
     await rest_client.wait_for_transaction(txn_hash)
 
     # :!:>section_5
-    txn_hash = await rest_client.create_token(
+    txn_hash = await token_client.create_token(
         alice,
         collection_name,
         token_name,
@@ -60,17 +61,19 @@ async def main():
     await rest_client.wait_for_transaction(txn_hash)
 
     # :!:>section_6
-    collection_data = await rest_client.get_collection(alice.address(), collection_name)
+    collection_data = await token_client.get_collection(
+        alice.address(), collection_name
+    )
     print(
         f"Alice's collection: {json.dumps(collection_data, indent=4, sort_keys=True)}"
     )  # <:!:section_6
     # :!:>section_7
-    balance = await rest_client.get_token_balance(
+    balance = await token_client.get_token_balance(
         alice.address(), alice.address(), collection_name, token_name, property_version
     )
     print(f"Alice's token balance: {balance}")  # <:!:section_7
     # :!:>section_8
-    token_data = await rest_client.get_token_data(
+    token_data = await token_client.get_token_data(
         alice.address(), collection_name, token_name, property_version
     )
     print(
@@ -79,7 +82,7 @@ async def main():
 
     print("\n=== Transferring the token to Bob ===")
     # :!:>section_9
-    txn_hash = await rest_client.offer_token(
+    txn_hash = await token_client.offer_token(
         alice,
         bob.address(),
         alice.address(),
@@ -91,7 +94,7 @@ async def main():
     await rest_client.wait_for_transaction(txn_hash)
 
     # :!:>section_10
-    txn_hash = await rest_client.claim_token(
+    txn_hash = await token_client.claim_token(
         bob,
         alice.address(),
         alice.address(),
@@ -101,10 +104,10 @@ async def main():
     )  # <:!:section_10
     await rest_client.wait_for_transaction(txn_hash)
 
-    alice_balance = rest_client.get_token_balance(
+    alice_balance = token_client.get_token_balance(
         alice.address(), alice.address(), collection_name, token_name, property_version
     )
-    bob_balance = rest_client.get_token_balance(
+    bob_balance = token_client.get_token_balance(
         bob.address(), alice.address(), collection_name, token_name, property_version
     )
     [alice_balance, bob_balance] = await asyncio.gather(*[alice_balance, bob_balance])
@@ -112,16 +115,15 @@ async def main():
     print(f"Bob's token balance: {bob_balance}")
 
     print("\n=== Transferring the token back to Alice using MultiAgent ===")
-    # :!:>section_11
-    txn_hash = await rest_client.direct_transfer_token(
+    txn_hash = await token_client.direct_transfer_token(
         bob, alice, alice.address(), collection_name, token_name, 0, 1
-    )  # <:!:section_11
+    )
     await rest_client.wait_for_transaction(txn_hash)
 
-    alice_balance = rest_client.get_token_balance(
+    alice_balance = token_client.get_token_balance(
         alice.address(), alice.address(), collection_name, token_name, property_version
     )
-    bob_balance = rest_client.get_token_balance(
+    bob_balance = token_client.get_token_balance(
         bob.address(), alice.address(), collection_name, token_name, property_version
     )
     [alice_balance, bob_balance] = await asyncio.gather(*[alice_balance, bob_balance])
